parse file substitutions
For platforms (such as OS X) where there is no support for response files,
or whenever response files cannot be nested, it is useful for command
line arguments to come from a file.  This patch adds a new syntax
$@varname or ${@varname} which treats the variable as a file name and
puts each line of the file into a separate command line argument.