graph: memoize expansion of command and rspfile_content
This will be necessary when command and rspfile_content will possibly
have file substitution in them, so that there is no possibility of
race conditions between multiple evaluations of the variables.

The file substitutions will be evaluated at most twice, once during
the dependency scan and once during the actual build process.